Added methods to distance matrix class that require both pairs and labels table.
Made pairs and table of distance matrix class properties instead of factory methods.

When merging distance matrices check that labels are equal.
If they are then pairs can be appended directly.
If they are not then pairs have to be converted for ids to labels and then added by looping.

Added test for merging distance matrices.

Fixes #10